requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.9.0
paho-mqtt==1.6.1
gunicorn==21.2.0
eventlet==0.33.3
Werkzeug==2.3.7